requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
ed25519==1.5
idna==3.4
numpy==1.24.3
protobuf==3.20.2
pyarrow==6.0.1
rai-sdk==0.6.14
requests==2.31.0
requests-toolbelt==1.0.0
urllib3==1.26.6
azure-storage-blob==12.17.0